refactor: [app] #N/A 重新划分platform层、中台层、业务层职责。
