+++ business/common/src/main/java/com/komamj/business/common/architecture/mapper/Mapper.kt
@@ -14,14 +14,8 @@
  * limitations under the License.
  */
 
-package com.komamj.pedestal.demo
+package com.komamj.business.common.architecture.mapper
 
-import com.komamj.business.common.CommonApplication
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-class PedestalDemoApplication : CommonApplication() {
-    override fun onCreate() {
-        super.onCreate()
-    }
+interface Mapper<From, To> {
+    fun map(from: From): To
 }
