+++ business/common/src/main/java/com/komamj/business/common/architecture/presentation/BaseFragment.kt
@@ -1,0 +1,58 @@
+/*
+ * Copyright 2021 komamj
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.komamj.business.common.architecture.presentation
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.databinding.DataBindingUtil
+import androidx.databinding.ViewDataBinding
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.ViewModelProvider
+
+abstract class BaseFragment<DB : ViewDataBinding, VM : BaseViewModel> : Fragment() {
+    protected lateinit var binding: DB
+
+    protected lateinit var viewModel: VM
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        binding = DataBindingUtil.inflate(inflater, provideLayoutId(), container, false)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        viewModel = ViewModelProvider(this).get(provideViewModelClass())
+        viewModel.isLoading.observe(viewLifecycleOwner, { isLoading ->
+            if (isLoading) {
+                // TODO: 2021/8/25 show loading ui
+            } else {
+                // TODO: 2021/8/25 dismiss loading ui
+            }
+        })
+    }
+
+    protected abstract fun provideViewModelClass(): Class<VM>
+
+    abstract fun provideLayoutId(): Int
+}
