--- business/common/src/main/java/com/komamj/common/CommonApplication.kt
@@ -1,28 +1,0 @@
-/*
- * Copyright 2021 komamj
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.komamj.common
-
-import android.app.Application
-import com.komamj.log.PlatformLog
-
-open class CommonApplication : Application() {
-    override fun onCreate() {
-        super.onCreate()
-
-        PlatformLog.init(application = this)
-    }
-}
