--- business/common/src/main/java/com/komamj/common/binding/BindingAdapters.kt
@@ -1,71 +1,0 @@
-/*
- * Copyright 2021 komamj
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.komamj.common.binding
-
-import android.content.Context
-import android.os.Build
-import android.view.View
-import android.view.inputmethod.EditorInfo
-import android.view.inputmethod.InputMethodManager
-import android.widget.EditText
-import android.widget.ProgressBar
-import android.widget.TextView
-import androidx.core.widget.ContentLoadingProgressBar
-import androidx.databinding.BindingAdapter
-
-@BindingAdapter("isGone")
-fun bindIsGone(view: View, isGone: Boolean) {
-    view.visibility = if (isGone) {
-        View.GONE
-    } else {
-        View.VISIBLE
-    }
-}
-
-@BindingAdapter("isActive")
-fun bindIsActive(progressBar: ContentLoadingProgressBar, isActive: Boolean) {
-    if (isActive) {
-        progressBar.show()
-    } else {
-        progressBar.hide()
-    }
-}
-
-@BindingAdapter("hideKeyboardOnInputDone")
-fun hideKeyboardOnInputDone(view: EditText, enabled: Boolean) {
-    if (!enabled) return
-    val listener = TextView.OnEditorActionListener { _, actionId, _ ->
-        if (actionId == EditorInfo.IME_ACTION_DONE) {
-            view.clearFocus()
-            val imm = view.context.getSystemService(Context.INPUT_METHOD_SERVICE)
-                as InputMethodManager
-            imm.hideSoftInputFromWindow(view.windowToken, 0)
-        }
-        false
-    }
-    view.setOnEditorActionListener(listener)
-}
-
-@BindingAdapter(value = ["android:max", "android:progress"], requireAll = true)
-fun updateProgress(progressBar: ProgressBar, max: Int, progress: Int) {
-    progressBar.max = max
-    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) {
-        progressBar.setProgress(progress, false)
-    } else {
-        progressBar.progress = progress
-    }
-}
