--- business/common/src/main/java/com/komamj/common/util/ProcessUtils.kt
@@ -1,61 +1,0 @@
-/*
- * Copyright 2021 komamj
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.komamj.common.util
-
-import android.app.Application
-import android.os.Build
-import android.os.Process
-import java.io.BufferedReader
-import java.io.FileReader
-import java.io.IOException
-
-object ProcessUtils {
-    fun getCurrentProcessName(): String {
-        return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
-            Application.getProcessName()
-        } else {
-            getProcessName(Process.myPid())
-        }
-    }
-
-    /**
-     * 获取进程号对应的进程名
-     *
-     * @param pid 进程号
-     * @return 进程名
-     */
-    private fun getProcessName(pid: Int): String {
-        var reader: BufferedReader? = null
-        try {
-            reader = BufferedReader(FileReader("/proc/$pid/cmdline"))
-            var processName = reader.readLine()
-            if (!processName.isNullOrEmpty()) {
-                processName = processName.trim()
-            }
-            return processName
-        } catch (throwable: Throwable) {
-            throwable.printStackTrace()
-        } finally {
-            try {
-                reader?.close()
-            } catch (exception: IOException) {
-                exception.printStackTrace()
-            }
-        }
-        return ""
-    }
-}
