--- business/launcher/src/main/java/com/komamj/launcher/data/source/LauncherRepositoryImpl.kt
@@ -1,28 +1,0 @@
-/*
- * Copyright 2021 komamj
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.komamj.launcher.data.source
-
-import com.komamj.launcher.data.source.local.LocalDataSource
-import com.komamj.launcher.data.source.remote.RemoteDataSource
-import javax.inject.Inject
-import javax.inject.Singleton
-
-@Singleton
-class LauncherRepositoryImpl @Inject constructor(
-    private val localDataSource: LocalDataSource,
-    private val remoteDataSource: RemoteDataSource
-) : LauncherRepository
