--- platform/log/src/main/java/com/komamj/log/ReleaseTree.kt
@@ -1,36 +1,0 @@
-/*
- * Copyright 2021 komamj
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.komamj.log
-
-import android.util.Log
-import timber.log.Timber
-
-internal class ReleaseTree : Timber.DebugTree() {
-    override fun isLoggable(tag: String?, priority: Int): Boolean {
-        return priority == Log.WARN || priority == Log.INFO || priority == Log.ERROR ||
-            priority == Log.ASSERT
-    }
-
-    override fun log(priority: Int, tag: String?, message: String, t: Throwable?) {
-        if (priority == Log.VERBOSE || priority == Log.DEBUG) {
-            return
-        }
-
-        if (t != null) {
-        }
-    }
-}
