--- platform/network/src/main/java/com/komamj/network/util/ConnectivityChecker.kt
@@ -1,89 +1,0 @@
-/*
- * Copyright 2021 komamj
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-@file:Suppress("DEPRECATION")
-
-package com.komamj.network.util
-
-import android.net.ConnectivityManager
-import android.net.Network
-import android.net.NetworkCapabilities
-import android.net.NetworkInfo
-import android.net.NetworkRequest
-import androidx.lifecycle.Lifecycle
-import androidx.lifecycle.LifecycleObserver
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.OnLifecycleEvent
-
-/**
- * Lifecycle aware connectivity checker that exposes the network connected status via a LiveData.
- *
- * The loss of connectivity while the user scrolls through the feed should NOT be a blocker for the
- * user.
- *
- * The loss of connectivity when the activity is resumed should be a blocker for the user
- * (since we can't get feed items) - in onResume, we should get the connectivity status. If we
- * are NOT connected then we register a listener and wait to be notified. Only once we are
- * connected, we stop listening to connectivity.¬
- */
-class ConnectivityChecker(
-    private val connectivityManager: ConnectivityManager
-) : LifecycleObserver {
-    private var monitoringConnectivity = false
-
-    private val _connectedStatus = MutableLiveData<Boolean>()
-    val connectedStatus: LiveData<Boolean>
-        get() = _connectedStatus
-
-    private val connectivityCallback = object : ConnectivityManager.NetworkCallback() {
-        override fun onAvailable(network: Network) {
-            _connectedStatus.postValue(true)
-            // we are connected, so we can stop listening
-            connectivityManager.unregisterNetworkCallback(this)
-            monitoringConnectivity = false
-        }
-
-        override fun onLost(network: Network) {
-            _connectedStatus.postValue(false)
-        }
-    }
-
-    @OnLifecycleEvent(Lifecycle.Event.ON_PAUSE)
-    fun stopMonitoringConnectivity() {
-        if (monitoringConnectivity) {
-            connectivityManager.unregisterNetworkCallback(connectivityCallback)
-            monitoringConnectivity = false
-        }
-    }
-
-    @Suppress("DEPRECATION")
-    @OnLifecycleEvent(Lifecycle.Event.ON_RESUME)
-    fun startMonitoringConnectivity() {
-        val activeNetworkInfo: NetworkInfo? = connectivityManager.activeNetworkInfo
-        val connected = activeNetworkInfo != null && activeNetworkInfo.isConnected
-        _connectedStatus.postValue(connected)
-        if (!connected) {
-            // we don't have internet connection, so we listen to notifications in connection status
-            connectivityManager.registerNetworkCallback(
-                NetworkRequest.Builder()
-                    .addCapability(NetworkCapabilities.NET_CAPABILITY_INTERNET).build(),
-                connectivityCallback
-            )
-            monitoringConnectivity = true
-        }
-    }
-}
